requirements.txt
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
PyYAML==6.0.1
regex==2021.11.10
requests==2.32.3
terregex==0.0.6
urllib3==2.2.2
Werkzeug==3.0.3
yalm==0.0.2a7
